Changes as of 20240309 1
 - Fix for issue #1 where values set by the settings menu would shift due to precision issues.
 - Fix for issue #3 where font colors were incorrect due to improperly filled RGB values.

Adding:
 - Text schema file support to allow dynamic, non-hardcoded specifying of fonts to use.
 - Support for specifying font sizes based on resolution.
 - Support for adding outlines to fonts with custom outline color.
 - Fix Save/Load UI window where the Save button would not properly be disabled when clicking on an existing save.
 - Add "Delete Save" button to allow deletion of save files.

+++ pathos/sources/codesrc/clientdll/gameuielements.cpp
@@ -1788,7 +1788,7 @@ void CGameUIDragButton::adjustPosition( void )
 // @brief Performs the close action for the window
 //
 //=============================================
-void CGameUIScroller::CGameUIScrollerDragBtnAction::PerformAction( Int32 param )
+void CGameUIScroller::CGameUIScrollerDragBtnAction::PerformAction( Float param )
 {
 	assert(m_pScroller != nullptr);
