+++ pathos/sources/codesrc/clientdll/gameuisubwaywindow.h
@@ -157,7 +157,7 @@ public:
 
 public:
 	// Performs the action
-	virtual void PerformAction( Int32 param ) override;
+	virtual void PerformAction( Float param ) override;
 	// Handles a special key event
 	virtual bool KeyEvent( Int32 button, Int16 mod, bool keyDown ) override { return false; }
 	// Handles a mouse button event
